backend/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
fastapi>=0.100.0
strawberry-graphql[fastapi]>=0.200.0
uvicorn>=0.22.0
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-multipart>=0.0.6
peft>=0.4.0
bitsandbytes>=0.41.0
accelerate>=0.20.0